Moves details-step/not selecting shipping method scenarios into dedicated file
